# CP-SAT Java: take advantage of cached constants in CpModel.indexFromConstant()